requirements.txt
# ===============================
# === LangChain + Chroma Stack
# ===============================
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain_community==0.3.23
langchain-text-splitters==0.3.8

chromadb>=1.0.0

# ===============================
# === OpenAI + Embeddings
# ===============================
openai==1.77.0

# ===============================
# === PDF & Document Parsing
# ===============================
unstructured==0.17.2
unstructured-inference==0.8.10
pikepdf==9.0.0
pdfminer.six==20231228
tika==2.6.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4
PyPDF2==3.0.1
langdetect==1.0.9
olefile==0.46                      # Prevent legacy version with Python 2 syntax

# ===============================
# === OCR and Table Support
# ===============================
pdf2image==1.17.0
pytesseract==0.3.13
timm==1.0.15

# ===============================
# === Torch + Numpy (Critical)
# ===============================
torch==2.1.1
torchvision==0.16.1
numpy<2                             # Avoid breaking torch compatibility

# ===============================
# === Transformers + Embeddings
# ===============================
transformers==4.51.3
sentence-transformers==3.0.1

# ===============================
# === Networking / Utilities
# ===============================
requests==2.28.2
httpx==0.28.1
httpx-sse==0.4.0
aiofiles==24.1.0
python-dotenv==1.1.0
tenacity==8.5.0

# ===============================
# === Database
# ===============================
sqlalchemy==2.0.40
psycopg2-binary==2.9.9

# ===============================
# === System-Level Dependencies
# ===============================
pi-heif==0.11.0                     # For HEIF/HEIC image decoding

# ===============================
# === Rate Limiting
# ===============================
ratelimit==2.2.1

# ===============================
# === Optional: FastAPI Runtime
# ===============================
fastapi==0.110.3
uvicorn[standard]==0.29.0
jinja2==3.1.3  # for HTML templating
